data-producers/pyproject.toml
[project]
name = "data-producers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.42.69",
    "confluent-kafka>=2.13.2",
    "fastapi>=0.135.1",
    "pydantic>=2.12.5",
    "uvicorn>=0.41.0",
    "python-dotenv"
]
